getblockchaininfo size_on_disk must return an integer

See merge request bitcoinrh/BRhodiumNode!49

+++ src/BRhodium.Bitcoin.Features.BlockStore/Models/GetBlockChainInfoModel.cs
@@ -32,7 +32,7 @@ namespace BRhodium.Bitcoin.Features.BlockStore.Models
         public string ChainWork { get; set; }
 
         [JsonProperty(PropertyName = "size_on_disk")]
-        public double SizeOnDisk { get; set; }
+        public int SizeOnDisk { get; set; }
 
         [JsonProperty(PropertyName = "pruned")]
         public bool Pruned { get; set; }
